feat: create base of the lib
Use a formContext with data manipulation in which
the field names and values will be related into a
schema object. Also, a hook `useForm` is created.

The types of the components are separated, such as
each component in its own folder inside the architecture.

All of this work shall be considered Work-in-progress (WIP).